+++ Sources/Present/Auth/SignUp/ViewController/UserProfileViewController.swift
@@ -1,6 +1,6 @@
 import UIKit
 
-final class UserInformationViewController: BaseVC<UserInformationViewModel> {
+final class UserProfileViewController: BaseVC<UserInformationViewModel> {
     private let profileImageView = UIImageView().then {
         $0.clipsToBounds = true
         $0.layer.cornerRadius = 50
